pyproject.toml
[tool.poetry]
name = "django-flutter-school"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "django_flutter_school"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
django-cors-headers = "^3.14.0"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
djangorestframework-simplejwt = "^5.2.2"
coverage = "^7.2.2"
pytest = "^7.2.2"
pytest-django = "^4.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
